refactor(aicore/filtering): split filters.py by concern
Address final review comment: 'filters.py file is too bloated. We should
separate concerns, like create a config.py and put the env var fetching
code, and models.py for the classes.'

Match the sibling-module convention (agent_memory, destination, etc.):

- _models.py: Severity enum + provider classes (ContentFilter,
  AzureContentFilter, LlamaGuard38bFilter) + direction containers
  (InputFiltering, OutputFiltering, ContentFiltering). No behaviour
  beyond construction and to_dict() serialisation.
- config.py: load_from_env() function + private _read_env_* helpers.
  Replaces ContentFiltering.from_env() classmethod with a module-level
  factory, matching agent_memory/config.py's load_from_env_or_mount()
  shape.
- _patch.py: FilteringOrchestrationConfig, _install, _active_cfg,
  _ORIGINAL_CONFIG — the LiteLLM transport monkeypatch.
- _api.py: set_filtering, disable_filtering, extract_filter_blocked
  decorated with @record_metrics. Honours the earlier review #12
  ('not good pattern to keep logic on init').
- __init__.py: thin re-export surface (the 12 public names) +
  module-layout doc.

filters.py deleted. Test imports updated to the new module paths;
mock paths in test_patch.py point at _patch.GenAIHubOrchestrationConfig.
ContentFiltering.from_env() classmethod removed in favour of
load_from_env() — callers (tests + _api.py) updated accordingly.
